Hook the frontend up to the backend
Adds a submit handler and some state variables to hold our values, as
well as a display for the value we got back.

The state variables hold the user's input (only lightly validated so
far with a NumberField) and the result we got back from the backend.

Our submit handler calls our conversion service and passes the user's
input, handling errors with a console log for now.