Make environment configurable
Added a production, development, and test environment file. Made the
convertRomanNumerals service use env vars from the new environment
files. Added a test that checks that the values from the environments
are used appropriately.

Added a test to make sure that we're sending the query parameter to the
backend.

Tweaked the vite config to look for the new env files.