Implement a range-checking input validating pipe
This pipe will implement range checking validation for us. The pipe
takes a couple of parameters that set the upper and lower bounds of the
range it validates, then throws an exception if the number falls outside
the range, and otherwise passes the number through if it falls within
the range.

Includes a bunch of tests that validate that it works as intended.