Incorporate input validation pipes, flesh out E2E tests
Add some validation pipes to our romannumeral endpoint, along with a
descriptive comment explaining how it works.

The pipes will validate that the query parameter is set, is an integer,
and that it is within 1-3999, using our new IntRagnePipe.

Reformats the E2Es to match our code style, and adds a bunch of new
tests that validate that we get back either a 400 or 200 status back
depending on if the input is invalid or valid, respectively.

The tests are not yet validating for correct roman numeral conversion,
since we haven't implemented that functionality.