Get tracing working
Adds tracing to the backend, using OpenTelemetry. Uses SigNoz locally to
gather that information. Traces are handled with the new tracer file
that lives in the main `src` directory. Information about how to
instrument that was gathered from a guide written by SigNoz, but also
cross-referenced with a number of other similar documents on the
internet.

Includes some instructions on how to get things up and running for local
metrics visualization.

Also put the new environment var into the compose file.