Switch logging to Winston, get logs attached to traces
The built-in NestJS logger isn't easily capable of integrating nicely
with OTEL (at least, not without a lot of my own leg work). As such, one
of the go-to options appears to be Winston, so switching to that. It was
pretty easy to integrate with nest via the `nest-winston` package, and
once I got that put in, the auto instrumentation in the OTEL library
just started picking the logs up. After a little research, that appears
to be the way it's supposed to work.

Also switched out the logger on the RomanNumeralsController to use the
new winston logger.