Fix unit tests after adding Winston
Forgot to come back and mock the Winston module for the tests.

Included the @golevelup/ts-jest package to make mocking it a little
easier. I've used this package quite a bit in the past and like the
utilities it adds. This tweak will make it so we can just auto-mock the
whole logging service without having to do something like add a
`jest.fn()` for each method on the logger if we accidentally add a call
to one later.