Flesh out the RomanNumeralsController and tests
Now that we have a model object that will be representing our return
value, we can start to flesh the controller out some more. We also have
our converter service, and even though it doesn't yet do anything, we
can call it, which enables us to implement better unit tests.

Adds some unit tests that mock the actual converter service and verify
that the controller constructs the result object properly.

Modifies the controller to inject the service via the constructor, and
call the service in the GET method. Also modifies said GET method to
properly construct our output, and changed the signature accordingly.

Adds our new service to the providers list on the RomanNumerals module.

Adds a few more E2E tests to ensure that the response comes out of the
service in the correct structure.