Modify frontend to expect ConversionResult DTO object
Since we just changed the backend to send back our properly formatted
object, we need the frontend to expect it.

Adds the interface for that object to a model folder on the frontend,
along with a comment about keeping them in sync (see diff), and adjusts
the service to expect the new model object.

Also adjusts said service's unit tests accordingly.