Perform roman numeral conversion
Add a bunch of unit tests to verify that we're correctly converting
roman numerals. We could be more comprehensive with this, but I'm not
going to add that for now. I'll come back and do that in a subsequent
pass.

Flesh out the actual conversion the rest of the way as well. Left a
number of comments explaining the methodology, so not repeating that
here. See diff.

Also added another e2e test that verifies that we actually send an
output correctly all the way through. I don't feel a strong need to
further test the actual value of the output, since that should be
handled by lower-level tests according to the testing pyramid
philosophy.