Flesh out testing and support for getDigits utility
Thought of a few cases that could end up getting passed to a utility
function like this in a hypothetical future and thought I should handle
them, since it was fairly easy.

Note that this isn't a perfect solution; other possibilities I
considered were to respond with an error, or make the representation of
the returned digits handle more cases. I didn't feel that either of
those were particularly warranted in this case, since, as currently
written, neither of these should end up getting passed here anyway, and
if in the future they are, a doc comment should be enough to warn a
user, and other, larger chunks of this code would need be modified to
handle them.