Handle range errors from conversion service
In the off chance that we somehow get through the validators on our
input query parameter, handle errors that come from the converter and
tell us that we have too many digits.

If it's a range error, translate that to a BadRequest, and if it's
something else, translate it to an InternalServerError.