Add a custom metric
This is mostly just to demonstrate how it could be done. The traces
functionality is already gathering metrics about how long our requests
are taking and what our responses are (i.e. error vs OK). This shows how
we could add another metric for custom things we wanted to measure.