Set max and min input for number field
Since we're using a number field here that provides us a minimum and
maximum value, let's use those. The backend will do error checking as
well, but this helps the user not enter the wrong thing.

Also add a hint to the user on the label what the acceptable values are,
since the number field just clamps the input to the acceptable range
automatically after the field loses focus.