Clean up boilerplate code, enable serving frontend on /
Clean out the AppController and AppService; they were created by the
template boilerplate and we don't need them in tihs project. Also
removed the generated e2e that was checking for the Hello World
response.

Removing these also made it possible for the static serve module to
provide the frontend at the root of the project, so `index.html` no
longer needs to be specified in the address bar when accessing the
frontend from the static serve location after building with the
container (or if building the frontend into the backend).

Also updated the main project README to reflect the changes.