Add loading circle
Sometimes (like when running in dev mode) the backend takes a few
seconds to load the results. In those cases, we should show a loading
circle to let the user know that we're working.

Include an aria label with it, because it doesn't have any descriptive
text otherwise associated with it.